backend/requirements.txt
# Server
fastapi==0.115.6
uvicorn[standard]==0.32.1
slowapi==0.1.9

# Transcript
youtube-transcript-api>=0.6.0

# LangChain & OpenAI (RAG)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
faiss-cpu>=1.7.0

# Env
python-dotenv>=1.0.0
